Collection interface, mock testing, placeholder for a deferred collection.
