+++ composer.json
@@ -15,7 +15,8 @@
         "codacy/coverage": "^1.4",
         "phan/phan": "^0.12.1",
         "infection/infection": "^0.8.1",
-        "phpstan/phpstan": "^0.9.2"
+        "phpstan/phpstan": "^0.9.2",
+        "mockery/mockery": "^1.0"
     },
     "license": "Apache-2.0",
     "minimum-stability": "dev",
