+++ src/LazyLINQ/LazyCollection.php
@@ -19,513 +19,6 @@ declare(strict_types=1);
 
 namespace LazyLINQ;
 
-use LazyLINQ\Errors\InvalidOperationException;
-
-/**
- * A partial lazy-only port of LINQ for PHP.
- */
-class LazyCollection extends \Pipeline\Simple implements Interfaces\Collection
+class LazyCollection extends Collection
 {
-    public static function from($source, ...$args)
-    {
-        if (is_array($source)) {
-            return new static(new \ArrayIterator($source));
-        }
-
-        if ($source instanceof \Traversable) {
-            return new static($source);
-        }
-
-        if ($source instanceof \Closure) {
-            return static::from($source(...$args));
-        }
-
-        return new static(new \ArrayIterator([$source]));
-    }
-
-    private function replace(callable $func)
-    {
-        return $this->map(static::from($func, clone $this));
-    }
-
-    public function aggregate($seed, callable $func, callable $resultSelector = null)
-    {
-        if ($resultSelector) {
-            return $resultSelector($this->reduce($func, $seed));
-        }
-
-        return $this->reduce($func, $seed);
-    }
-
-    public function all(callable $predicate = null): bool
-    {
-        if (!$predicate) {
-            return $this->allTrue();
-        }
-
-        foreach ($this as $value) {
-            if (!$predicate($value)) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    private function allTrue()
-    {
-        foreach ($this as $value) {
-            if (!$value) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    public function any(callable $predicate = null): bool
-    {
-        if ($predicate) {
-            $this->filter($predicate);
-        }
-
-        /*
-         * foreach is marginally faster than using embedded \Iterator:
-         *
-         * $this->getIterator()->rewind();
-         * return $this->getIterator()->valid();
-         */
-        foreach ($this as $value) {
-            return true;
-        }
-
-        return false;
-    }
-
-    public function append($element)
-    {
-        // `yield from` is about four times faster than \AppendIterator
-        // and about 50% faster than `foreach-yield`
-        return $this->replace(static function ($previous) use ($element) {
-            yield from $previous;
-            yield $element;
-        });
-    }
-
-    public function average(callable $selector = null)
-    {
-        if ($selector) {
-            $this->map($selector);
-        }
-
-        $result = $this->reduce(static function ($carry, $value) {
-            $carry->sum += $value;
-            $carry->count += 1;
-
-            return $carry;
-        }, (object) ['sum' => 0, 'count' => 0]);
-
-        return $result->sum / $result->count;
-    }
-
-    public function cast($type)
-    {
-        return $this->map(static function ($value) use ($type) {
-            if (settype($value, $type)) {
-                yield $value;
-            }
-        });
-    }
-
-    /**
-     * Concatenates two sequences.
-     *
-     * @param \Traversable|array $second the sequence to concatenate
-     *
-     * @return static new instance
-     */
-    public function concat($second)
-    {
-        return $this->replace(static function ($previous) use ($second) {
-            yield from $previous;
-            yield from $second;
-        });
-    }
-
-    public function contains($value, callable $comparer = null)
-    {
-        if (!$comparer) {
-            return $this->containsAny($value);
-        }
-
-        foreach ($this as $sample) {
-            if ($comparer($sample, $value)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private function containsAny($value)
-    {
-        // We could refactor this with a `map()` call, but that's an extra function call for each iteration.
-        foreach ($this as $sample) {
-            if ($sample == $value) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    public function count(callable $predicate = null)
-    {
-        if ($predicate) {
-            $this->map($predicate)->filter();
-        }
-
-        return \iterator_count($this);
-    }
-
-    public function distinct(callable $comparer = null)
-    {
-        return $this->map(static function ($value) use ($comparer) {
-            static $previous;
-            static $previousSeen = false;
-
-            if (!$previousSeen) {
-                $previousSeen = true;
-                $previous = $value;
-                yield $value;
-            }
-
-            if ($comparer ? !$comparer($value, $previous) : $value != $previous) {
-                $previous = $value;
-                yield $value;
-            }
-        });
-    }
-
-    public static function empty()
-    {
-        return static::from([]);
-    }
-
-    public function except($collection, callable $comparer = null)
-    {
-        if (!$comparer && is_array($collection)) {
-            return $this->exceptArray($collection);
-        }
-
-        if (!$comparer) {
-            return $this->exceptEquals($collection);
-        }
-
-        return $this->replace(static function ($previous) use ($collection, $comparer) {
-            foreach ($previous as $value) {
-                foreach ($collection as $excluded) {
-                    if ($comparer($value, $excluded)) {
-                        continue 2;
-                    }
-                }
-
-                yield $value;
-            }
-        })->distinct();
-    }
-
-    private function exceptArray(array $collection)
-    {
-        return $this->replace(static function ($previous) use ($collection) {
-            foreach ($previous as $value) {
-                if (!in_array($value, $collection)) {
-                    yield $value;
-                }
-            }
-        })->distinct();
-    }
-
-    private function exceptEquals($collection)
-    {
-        return $this->replace(static function ($previous) use ($collection) {
-            foreach ($previous as $value) {
-                foreach ($collection as $excluded) {
-                    if ($value == $excluded) {
-                        continue 2;
-                    }
-                }
-
-                yield $value;
-            }
-        })->distinct();
-    }
-
-    public function first(callable $predicate = null)
-    {
-        if ($predicate) {
-            $this->filter($predicate);
-        }
-
-        foreach ($this as $value) {
-            return $value;
-        }
-    }
-
-    public function last(callable $predicate = null)
-    {
-        if ($predicate) {
-            $this->filter($predicate);
-        }
-
-        $value = null;
-
-        foreach ($this as $value) {
-            // Not casting to an array here to save memory and some cycles
-        }
-
-        return $value;
-    }
-
-    public function max(callable $selector = null)
-    {
-        if ($selector) {
-            $this->map($selector);
-        }
-
-        $max = null; // everything is greater than null
-
-        // We can load all values and be done with max(...$this),
-        // but all values could take more memory than we have
-        foreach ($this as $value) {
-            if ($value > $max) {
-                $max = $value;
-            }
-        }
-
-        return $max;
-    }
-
-    public function min(callable $selector = null)
-    {
-        if ($selector) {
-            $this->map($selector);
-        }
-
-        $min = null;
-
-        // We can load all values and be done with min(...$this),
-        // but all values could take more memory than we have
-        foreach ($this as $value) {
-            $min = $value;
-            break;
-        }
-
-        foreach ($this as $value) {
-            if ($value < $min) {
-                $min = $value;
-            }
-        }
-
-        return $min;
-    }
-
-    public function ofType($type)
-    {
-        return $this->filter(static function ($value) use ($type) {
-            return gettype($value) == $type;
-        });
-    }
-
-    public function ofClass($className)
-    {
-        return $this->filter(static function ($value) use ($className) {
-            return is_object($value) && get_class($value) == $className;
-        });
-    }
-
-    public function prepend($element)
-    {
-        return $this->replace(static function ($previous) use ($element) {
-            yield $element;
-            yield from $previous;
-        });
-    }
-
-    /**
-     * Determines from which number of sequential integers a lazy generator should be used.
-     *
-     * @var int
-     */
-    const LAZY_RANGE_MIN_COUNT = 101;
-
-    public static function range(int $start, int $count)
-    {
-        /*
-         * Typical memory usage is the following:
-         *
-         * On 100 ints: 8432 with range(), 5232 with a generator.
-         * On 10000 ints: 528624 with range(), 5232 with a generator.
-         */
-
-        if ($count < static::LAZY_RANGE_MIN_COUNT) {
-            return new static(new \ArrayIterator(range($start, $start + $count - 1)));
-        }
-
-        return static::from(static function () use ($start, $count) {
-            do {
-                yield $start;
-                $start += 1;
-            } while ($count -= 1);
-        });
-    }
-
-    public static function repeat($element, int $count)
-    {
-        return static::from(static function () use ($element, $count) {
-            do {
-                yield $element;
-            } while ($count -= 1);
-        });
-    }
-
-    public function select(callable $selector)
-    {
-        return $this->map($selector);
-    }
-
-    public function selectMany(callable $selector = null)
-    {
-        if ($selector) {
-            $this->map($selector);
-        }
-
-        return $this->unpack();
-    }
-
-    public function single(callable $predicate = null)
-    {
-        if ($predicate) {
-            $this->filter($predicate);
-        }
-
-        $found = false;
-        $foundValue = null;
-
-        foreach ($this as $value) {
-            if ($found) {
-                throw new InvalidOperationException();
-            }
-
-            $found = true;
-            $foundValue = $value;
-        }
-
-        return $foundValue;
-    }
-
-    public function skip(int $count)
-    {
-        return $this->filter(static function () use ($count) {
-            static $skipped = 0;
-            $skipped += 1;
-
-            return $skipped > $count;
-        });
-    }
-
-    public function skipWhile(callable $predicate)
-    {
-        return $this->filter(static function ($value) use ($predicate) {
-            static $bypass = true;
-
-            if (!$bypass) {
-                return true;
-            }
-
-            if (!$bypass = $predicate($value)) {
-                return true;
-            }
-
-            return false;
-        });
-    }
-
-    public function sum(callable $selector = null)
-    {
-        if ($selector) {
-            $this->map($selector);
-        }
-
-        return $this->reduce();
-    }
-
-    public function take(int $count)
-    {
-        return $this->replace(static function ($previous) use ($count) {
-            foreach ($previous as $value) {
-                if ($count <= 0) {
-                    break;
-                }
-
-                yield $value;
-
-                $count -= 1;
-            }
-        });
-    }
-
-    public function takeWhile(callable $predicate)
-    {
-        return $this->replace(static function ($previous) use ($predicate) {
-            foreach ($previous as $value) {
-                if (!$predicate($value)) {
-                    break;
-                }
-
-                yield $value;
-            }
-        });
-    }
-
-    public function where(callable $predicate)
-    {
-        return $this->filter($predicate);
-    }
-
-    public function zip($collection, callable $resultSelector = null)
-    {
-        // Collection must be non-rewindable. A generator can't be used here.
-        // \NoRewindIterator needs \Iterator, not \IteratorAggregate
-        $collection = new \NoRewindIterator(
-            $collection instanceof \IteratorAggregate
-            ? $collection->getIterator()
-            : static::from($collection)->getIterator()
-        );
-
-        $this->replace(static function ($previous) use ($collection) {
-            foreach ($previous as $firstValue) {
-                foreach ($collection as $secondValue) {
-                    yield [
-                        $firstValue,
-                        $secondValue,
-                    ];
-
-                    $collection->next();
-                    break;
-                }
-            }
-        });
-
-        if ($resultSelector) {
-            $this->unpack($resultSelector);
-        }
-
-        return $this;
-    }
-
-    public function jsonSerialize()
-    {
-        return $this->toArray();
-    }
 }
